pyproject.toml
[project]
name = "guardrail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "nltk>=3.9.1",
    "pymongo>=4.13.2",
    "ruff>=0.12.0",
    "scikit-learn>=1.7.0",
    "sentence-transformers>=4.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
